+++ server/standalone/interceptors/provider_interceptor_test.go
@@ -39,6 +39,14 @@ func (p *mockTransactionalDBALProvider) GetTransactionalDBAL(_ context.Context, 
 	return p.m, nil
 }
 
+type mockHealthReporter struct {
+	mock.Mock
+}
+
+func (m *mockHealthReporter) Shutdown() {
+	m.Called()
+}
+
 func TestExtractProvider(t *testing.T) {
 	ctx := context.TODO()
 
@@ -57,12 +65,15 @@ func TestExtractProvider(t *testing.T) {
 
 func TestInjectProvider(t *testing.T) {
 	publisher := new(common.MockPublisher)
+	publisher.On("IsReady").Return(true)
 
 	db := new(mockedTransactionDBAL)
 	db.On("Commit").Once().Return(nil)
 	dbProvider := &mockTransactionalDBALProvider{db}
 
-	interceptor := NewProviderInterceptor(dbProvider, publisher)
+	healthcheck := new(mockHealthReporter)
+
+	interceptor := NewProviderInterceptor(dbProvider, publisher, healthcheck)
 
 	unaryInfo := &grpc.UnaryServerInfo{
 		FullMethod: "TestService.UnaryMethod",
@@ -83,8 +94,10 @@ func TestInjectProvider(t *testing.T) {
 
 func TestOnSuccess(t *testing.T) {
 	publisher := new(common.MockPublisher)
+	publisher.On("IsReady").Return(true)
 	db := new(mockedTransactionDBAL)
 	dbProvider := &mockTransactionalDBALProvider{db}
+	healthcheck := new(mockHealthReporter)
 
 	db.On("Commit").Once().Return(nil)
 
@@ -95,7 +108,7 @@ func TestOnSuccess(t *testing.T) {
 		wg.Done()
 	})
 
-	interceptor := NewProviderInterceptor(dbProvider, publisher)
+	interceptor := NewProviderInterceptor(dbProvider, publisher, healthcheck)
 
 	unaryInfo := &grpc.UnaryServerInfo{
 		FullMethod: "TestService.UnaryMethod",
@@ -123,12 +136,14 @@ func TestOnSuccess(t *testing.T) {
 
 func TestOnError(t *testing.T) {
 	publisher := new(common.MockPublisher)
+	publisher.On("IsReady").Return(true)
 	db := new(mockedTransactionDBAL)
 	dbProvider := &mockTransactionalDBALProvider{db}
+	healthcheck := new(mockHealthReporter)
 
 	db.On("Rollback").Once().Return(nil)
 
-	interceptor := NewProviderInterceptor(dbProvider, publisher)
+	interceptor := NewProviderInterceptor(dbProvider, publisher, healthcheck)
 
 	unaryInfo := &grpc.UnaryServerInfo{
 		FullMethod: "TestService.UnaryMethod",
@@ -150,4 +165,32 @@ func TestOnError(t *testing.T) {
 
 	db.AssertExpectations(t)
 	publisher.AssertExpectations(t)
+}
+
+func TestStopServingOnBrokerNotReady(t *testing.T) {
+	publisher := new(common.MockPublisher)
+	publisher.On("IsReady").Return(false)
+	db := new(mockedTransactionDBAL)
+	dbProvider := &mockTransactionalDBALProvider{db}
+	healthcheck := new(mockHealthReporter)
+	healthcheck.On("Shutdown")
+
+	interceptor := NewProviderInterceptor(dbProvider, publisher, healthcheck)
+
+	unaryInfo := &grpc.UnaryServerInfo{
+		FullMethod: "TestService.UnaryMethod",
+	}
+	unaryHandler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		assert.Fail(t, "handler should not be called")
+		return nil, errors.New("test error")
+	}
+
+	ctx := common.WithChannel(context.TODO(), "testChannel")
+	res, err := interceptor.Intercept(ctx, "test", unaryInfo, unaryHandler)
+	assert.Nil(t, res)
+	assert.Error(t, err)
+
+	db.AssertExpectations(t)
+	publisher.AssertExpectations(t)
+	healthcheck.AssertExpectations(t)
 }